[Chore] rename  : postCard -> rollingCard
