+++ src/components/RollingCard/RollingCard.jsx
@@ -1,9 +1,9 @@
-import * as S from "./PostCardUl.styles";
+import * as S from "./RollingCard.styles";
 import trash from "../../assets/icons/trash.svg";
 import cat from "../../assets/icons/cat.jpeg";
 import Badge from "../Badge/Badge";
 
-export default function PostCardUI({ data, isEdit = false, onClick }) {
+export default function RollingCard({ data, isEdit = false, onClick }) {
   const fonts = {
     "Noto Sans": "Noto Sans KR",
     Pretendard: "Pretendard",
@@ -12,9 +12,9 @@ export default function PostCardUI({ data, isEdit = false, onClick }) {
   };
 
   const date = new Date(data.createdAt);
-  const formattedDate = `${date.getFullYear()}.${String(
-    date.getMonth() + 1
-  ).padStart(2, "0")}.${String(date.getDate()).padStart(2, "0")}`;
+  const formattedDate = `${date.getFullYear()}.${String(date.getMonth() + 1).padStart(2, "0")}.${String(
+    date.getDate()
+  ).padStart(2, "0")}`;
 
   const fontFamily = fonts[data.font];
